Brick Breaker
This project aims to develop a microcontroller-based gaming machine using various tools. Players will attempt to break bricks on an OLED screen by deflecting a ball using a physical paddle control device. The player will use a physical "paddle control device," which will be controlled by a potentiometer. An OLED screen will display the game area to the players, with bricks at the top and a paddle controlled by the player and a bouncing ball at the bottom. The player's objective is to break the bricks at the top using the ball. The ball will bounce upwards using the paddle controlled by the player. When the ball hits a brick, the brick will disappear, and the player will gain "1" point. This score should be displayed using a 7-segment display. The direction of the ball will change when it hits the paddle or the walls. The right and left sides of the screen will be considered walls. If the game is successfully completed, the next game will start with the ball's speed increased by 20% compared to the previous one. The scoreboard screen should not reset when moving to the next level. The game will end if the ball falls below the paddle. The project will be developed using wok and Arduino Mega.